% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD.SCMapping.R
\name{assign_sc_cords}
\alias{assign_sc_cords}
\title{The function to assign the spatial location information for
each single cell}
\usage{
assign_sc_cords(mappint_spot_cell_cor, cords_new, numcell, sc_eset, ct_varname)
}
\arguments{
\item{mappint_spot_cell_cor}{a mapped correlation matrix indicating the
relashionship between each measured spatial location and the single cell in
the scRNAseq reference}

\item{cords_new}{output from the function get_high_res_cords}

\item{numcell}{a numeric value indicating the number of single cells in each
measured location, we suggest 20 for ST technology, 7 for 10x Viisum and 2
for Slide-seq}

\item{sc_eset}{a single cell experiment object stored in CARD object}

\item{ct_varname}{character, the name of the column in metaData that
specifies the cell type annotation information, stroed in CARD object}
}
\value{
Return the assigned spatial location information for the mapped
single cell
}
\description{
The function to assign the spatial location information for
each single cell
}
