% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerList.R
\docType{data}
\name{markerList}
\alias{markerList}
\title{marker gene list}
\format{
An object of class \code{list} of length 20.
}
\usage{
data(markerList)
}
\description{
The marker gene list is a list format with each element of the list being the
cell type specific gene markers.
}
\keyword{datasets}
