% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD.imputation.R
\name{norm_coords_train_test}
\alias{norm_coords_train_test}
\title{Normalize the new spatial locations without changing the shape and
relative positions}
\usage{
norm_coords_train_test(location_orig, train_ind, test_ind)
}
\arguments{
\item{location_orig}{Data frame, spatial location data frame of the original
spatial resolved transcriptomics dataset, stored in the
spatialCoords(CARD_object)}

\item{train_ind}{Vector, Index of the original spatial locations}

\item{test_ind}{Vector, Index of the newly grided spatial locations}
}
\value{
Return the normalized spatial location data frame
}
\description{
Normalize the new spatial locations without changing the shape and
relative positions
}
