% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD.imputation.R
\name{sample_grid_within}
\alias{sample_grid_within}
\title{Make new spatial locations on unmeasured tissue through grids.}
\usage{
sample_grid_within(location, num_sample, concavity = 2)
}
\arguments{
\item{location}{Data frame, spatial location data frame of the original
spatial resolved transcriptomics dataset, stored in the
spatialCoords(CARD_object)}

\item{num_sample}{Numeric, approximate number of cells in grid within the
shape of the spatial location data frame}

\item{concavity}{Numeric, a relative measure of concavity. The default is
2.0, which can prodecure detailed enough shapes. Infinity results in a
convex hull while 1 results in a more detailed shape.}
}
\value{
Return a list of data frame with newly grided points
}
\description{
Make new spatial locations on unmeasured tissue through grids.
}
