% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_count.R
\docType{data}
\name{sc_count}
\alias{sc_count}
\title{scRNA-seq count data}
\format{
An object of class \code{dgCMatrix} with 7000 rows and 1926 columns.
}
\usage{
data(sc_count)
}
\description{
The scRNA-seq count data must be in the format of matrix or sparseMatrix, while
each row represents a gene and each column represents a cell.
}
\keyword{datasets}
