% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_meta.R
\docType{data}
\name{sc_meta}
\alias{sc_meta}
\title{scRNAseq meta data}
\format{
An object of class \code{data.frame} with 1926 rows and 3 columns.
}
\usage{
data(sc_meta)
}
\description{
The scRNAseq meta data must be in the format of data frame while each row
represents a cell. The rownames of the scRNAseq meta data should match
exactly with the column names of the scRNAseq count data. The sc_meta data
must contain the column indicating the cell type assignment for each cell
(e.g., “cellType” column in the example sc_meta data). Sample/subject
information should be provided, if there is only one sample, we can add a
column by sc_meta$sampleInfo = "sample1".
}
\keyword{datasets}
