% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD.prop.R
\name{select_info}
\alias{select_info}
\title{Select Informative Genes used in the deconvolution}
\usage{
select_info(basis, sc_eset, commongene, ct_select, ct_varname)
}
\arguments{
\item{basis}{Reference basis matrix.}

\item{sc_eset}{scRNAseq data along with meta data stored in the S4 class
format (SingleCellExperiment).}

\item{commongene}{common genes between scRNAseq count data and spatial
resolved transcriptomics data.}

\item{ct_select}{vector of cell type names that you are interested in to
deconvolute, default as NULL. If NULL, then use all cell types provided by
single cell dataset;}

\item{ct_varname}{character, the name of the column in metaData that
specifies the cell type annotation information}
}
\value{
a vector of informative genes selected
}
\description{
Select Informative Genes used in the deconvolution
}
