% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{show,CARD-method}
\alias{show,CARD-method}
\title{Show method for the CARD class}
\usage{
\S4method{show}{CARD}(object)
}
\arguments{
\item{object}{An object of class \code{CARD}.}
}
\value{
A concise summary of the \code{CARD} object is printed to the console.
}
\description{
This method provides a concise summary of an object of class \code{CARD},
displaying key information including the project name, the number of spots,
the number of cell types, and a sample of the
\code{Proportion_CARD} matrix.
}
\examples{
data(spatial_count)
data(spatial_location)
data(sc_count)
data(sc_meta)
CARD_obj <- createCARDObject(
    sc_count = sc_count,
    sc_meta = sc_meta,
    spatial_count = spatial_count,
    spatial_location = spatial_location,
    ct.varname = "cellType",
    ct.select = unique(sc_meta$cellType),
    sample.varname = "sampleInfo",
    minCountGene = 100,
    minCountSpot = 5,
    sce = NULL,
    spe = NULL
)
show(CARD_obj)

}
