% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,CARDfree-method}
\alias{show,CARDfree-method}
\title{Show method for the CARDfree class}
\usage{
\S4method{show}{CARDfree}(object)
}
\arguments{
\item{object}{An object of class \code{CARDfree}.}
}
\value{
A concise summary of the \code{CARDfree} object is
printed to the console.
}
\description{
This method provides a concise summary of an object of
class \code{CARDfree}, displaying key information including the project
name, the number of spots, the number of cell types, and a sample of the
\code{Proportion_CARD} matrix.
}
