% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_count.R
\docType{data}
\name{spatial_count}
\alias{spatial_count}
\title{Spatial transcriptomics count data}
\format{
An object of class \code{dgCMatrix} with 11000 rows and 428 columns.
}
\usage{
data(spatial_count)
}
\description{
The spatial transcriptomics count data must be in the format of matrix or
sparseMatrix, while each row represents a gene and each column represents a
spatial location. The column names of the spatial data can be in the “XcoordxYcoord”
(i.e., 10x10) format, but you can also maintain your original spot names,
for example, barcode names.
}
\keyword{datasets}
