% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_location.R
\docType{data}
\name{spatial_location}
\alias{spatial_location}
\title{Spatial location data}
\format{
An object of class \code{data.frame} with 428 rows and 2 columns.
}
\usage{
data(spatial_location)
}
\description{
The spatial location data must be in the format of data frame while each row
represents a spatial location, the first column represents the x coordinate
and the second column represents the y coordinate. The rownames of the
spatial location data frame should match exactly with the column names of the
spatial_count.
}
\keyword{datasets}
