% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD.prop.R
\name{CARD_deconvolution}
\alias{CARD_deconvolution}
\title{Spatially Informed Cell Type Deconvolution for Spatial Transcriptomics
by CARD}
\usage{
CARD_deconvolution(CARD_object, iseed = 20200107)
}
\arguments{
\item{CARD_object}{CARD object create by the createCARDObject function}

\item{iseed}{Seed for initialization}
}
\value{
Returns a CARD object with estimated cell type proportion stored
in object@Proportion_CARD.
}
\description{
Spatially Informed Cell Type Deconvolution for Spatial Transcriptomics
by CARD
}
\examples{
library(RcppML)
library(NMF)
library(RcppArmadillo)
data(spatial_count)
data(spatial_location)
data(sc_count)
data(sc_meta)
CARD_obj <- createCARDObject(
    sc_count = sc_count,
    sc_meta = sc_meta,
    spatial_count = spatial_count,
    spatial_location = spatial_location,
    ct.varname = "cellType",
    ct.select = unique(sc_meta$cellType),
    sample.varname = "sampleInfo",
    minCountGene = 100,
    minCountSpot = 5
)
CARD_obj <- CARD_deconvolution(CARD_object = CARD_obj)
}
