% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,CARD-method}
\alias{show,CARD-method}
\title{Show method for the CARD class}
\usage{
\S4method{show}{CARD}(object)
}
\arguments{
\item{object}{An object of class \code{CARD}.}
}
\value{
A concise summary of the \code{CARD} object is printed to the
console.
}
\description{
This method provides a concise summary of an object of class \code{CARD},
displaying key information including the project name, the number of spots,
the number of cell types, and a sample of the
\code{Proportion_CARD} matrix.
}
