% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{show,CARDfree-method}
\alias{show,CARDfree-method}
\title{Show method for the CARDfree class}
\usage{
\S4method{show}{CARDfree}(object)
}
\arguments{
\item{object}{An object of class \code{CARDfree}.}
}
\value{
A concise summary of the \code{CARDfree} object is
printed to the console.
}
\description{
This method provides a concise summary of an object of class \code{CARDfree},
displaying key information including the project name, the number of spots,
the number of cell types, and a sample of the
\code{Proportion_CARD} matrix.
}
\examples{
data(markerList)
data(spatial_count)
data(spatial_location)
CARDfree_obj <- createCARDfreeObject(
    markerList = markerList,
    spatial_count = spatial_count,
    spatial_location = spatial_location,
    minCountGene = 100,
    minCountSpot = 5,
    spe = NULL
)
show(CARDfree_obj)

}
